fix: remove nested CDATA declarations in content (#231)
During OLX node creation, some XBlocks will wrap their entire contents
with CDATA (e.g. the HTMLBlock in libraries). This can cause errors
when that wrapped content contains its own CDATA sections, because
CDATA sections cannot nest.

This can happen in HTML documents where the contents of a <script> tag
are sometimes enclosed in CDATA. Modern HTML does not require this, but
it was a common practice when XHTML was an accepted standard, as this
would make those documents easier to parse as XML.

This commit will remove nested CDATA declarations while preserving
their content, so that the top-level CDATA can work as expected.