feat: Source specifying for relative links is allowed (#232) [FC-0063]
It is possible that an .imscc course dump contains files with relative
links to resources for some reason not included into it (e.g. some
LMS static files that are not processed by its exporting tool). It can
show up as broken images, missing styles etc.

To handle such situations, we now allow you to specify the source for
relative links during the conversion process using the -s flag:

  cc2olx -i <IMSCC_FILE> -s <RELATIVE_LINKS_SOURCE>