feat: implement PDF block processing (#253)
There situations when some custom XBlocks not included in the default
Open edX installation are installed on the target instance, and they
can display Common Cartridge content. For example, the PDF XBlock
allows displaying the PDF file directly on the course page.

This commit adds a --content_types_with_custom_blocks (or -c) CLI
argument to the converter to allow users to specify the XBlock types
that can be created if the appropriate Common Cartridge content
corresponds to them. The interface supports specifying multiple XBlock
types in the future, but only PDF type is supported at this time.